Rename: add hook api sidebar folder and move sidebar query
